Don't overflow original ids

+++ core/src/main/java/org/neo4j/graphalgo/core/utils/paged/SparseLongArray.java
@@ -162,7 +162,7 @@ public final class SparseLongArray {
                     mask >>= shift;
                 }
 
-                return (blockIndex << BLOCK_SHIFT) + pos;
+                return (((long) blockIndex) << BLOCK_SHIFT) + pos;
             }
             originalId += idsInPage;
         }
